reorganize: analytics view at root
