+++ apps/analytics/app/_components/EventsTable/Columns.tsx
