+++ apps/analytics/app/_components/EventsTable/MetadataDialog.tsx
