+++ apps/analytics/app/_components/RegionsTable/Columns.tsx
