+++ apps/analytics/app/page.tsx
@@ -1,18 +1,21 @@
-"use client";
-import { useState } from "react";
-import { ShieldAlert, BarChartBig, X, Menu } from "lucide-react";
-import { AnalyticsStats } from "./analytics/_components/AnalyticsStats";
-import EventsTable from "./analytics/_components/EventsTable/EventsTable";
-import Sidebar from "~/components/Sidebar";
-const navigation = [
-  { name: "Analytics", href: "#", icon: BarChartBig, current: true },
-  { name: "Errors", href: "#", icon: ShieldAlert, current: false },
-];
+import { Card, CardContent, CardHeader } from "~/components/ui/card";
+import { AnalyticsStats } from "./_components/AnalyticsStats";
+import EventsTable from "./_components/EventsTable/EventsTable";
+import RegionsTable from "./_components/RegionsTable/RegionsTable";
 
-function classNames(...classes: string[]) {
-  return classes.filter(Boolean).join(" ");
-}
-
-export default function Dashboard() {
-  return <div>main content</div>;
+export default function AnalyticsView() {
+  return (
+    <div className="space-y-4">
+      <AnalyticsStats />
+      <div className="grid gap-4 grid-cols-2">
+        <EventsTable />
+        <Card>
+          <CardHeader>Regions</CardHeader>
+          <CardContent>
+            <RegionsTable />
+          </CardContent>
+        </Card>
+      </div>
+    </div>
+  );
 }
